chore: add a basic test for the useSearchIndex hook
