+++ packages/api-reference/src/helpers/createEmptySpecification.ts
@@ -1,4 +1,5 @@
 import type { Spec } from '@scalar/oas-utils'
+import type { OpenAPI } from '@scalar/openapi-parser'
 
 import { deepMerge } from './deepMerge'
 
@@ -6,7 +7,9 @@ import { deepMerge } from './deepMerge'
  * Creates an empty specification object.
  * The returning object has the same structure as a valid OpenAPI specification, but everything is empty.
  */
-export function createEmptySpecification(partialSpecification?: Partial<Spec>) {
+export function createEmptySpecification(
+  partialSpecification?: Partial<OpenAPI.Document>,
+) {
   return deepMerge(partialSpecification ?? {}, {
     info: {
       title: '',
